+++ mycodo/databases/models/misc.py
@@ -35,6 +35,9 @@ class Misc(CRUDMixin, db.Model):
     mycodo_upgrade_available = db.Column(db.Boolean, default=False)  # Stores if an upgrade is available
     rpyc_timeout = db.Column(db.Integer, default=30)
     daemon_debug_mode = db.Column(db.Boolean, default=False)
+    net_test_ip = db.Column(db.String, default='8.8.8.8')
+    net_test_port = db.Column(db.Integer, default=53)
+    net_test_timeout = db.Column(db.Integer, default=3)
 
     def __repr__(self):
         return "<{cls}(id={s.id})>".format(s=self, cls=self.__class__.__name__)